Shows how logging effects the output while using munit-cats-effect
Mill test results don't appear in the shell output when the test is
run with info level logging (very verbose) turned on. When the logging is
set to error level logging (no errors occur in a normal test
executino) the test results are included in the output. When ran via
metals test command, the test results are included in the output.

- `output.txt` shows the test output when the test is run with no
logback-test.xml in place

- `mill-metals-output.txt` shows the output when the test is run
through metals lsp-dap test

- `mill-output-error-only-logging.txt` shows the output when the test
is run with logback-test.xml as it appears in the repository.